Support Synchronous Function Calls
Problem
-------
Behavioral programming was conceived as an event-driven paradigm in
which programs would remain running, receiving and emitting events.

However, it can also be used for composing behaviors in the same way as
a synchronous function call would be used.

For example, a number of business rules may be relevant to a single
function call. These can be composed in conventional ways, by logic
engines, etc.

But they can also be composed with BP.

Solution
--------
Add an `execute!` function that takes the bthreads and any events to
execute, submits the events, and returns the terminal event.