Move bthread prioritization into the bprograms
Problem
-------
Bthreads are only prioritized in the context of a program. The same
bthread may have a different priority in a different bprogram.

Further, for purposes of model checking, we can eliminate complexity by
requiring that all bthreads have a distinct order with respect to each
other.

Solution
--------
Bthread prioritization is done by passing an ordered sequence into the
function that makes the bprogram.