Checker takes a function that makes bthreads, instead of bthreads
Problem
-------
The checker needs to set up the bthreads more than once. For example,
it may check against all permutations of a collection of events.

But bthreads are stateful. Hence, if they are passed into the checker,
the bthreads will have the state from the first permutation when the
second permutation is tested.

Solution
--------
Instead of taking `:bthreads` as an argument, take a `:make-bthreads`
argument that can be called on each iteration.