Scenario test utility for pavlov
Problem
-------
Pavlov provides a model checker, which is very useful for testing whole
programs.

But sometimes it would be useful to declare a scenario (as a sequence of
events) and validate that is a possible execution path for a program.

When that execution path is not possible, it would be helpful to know

1. The step in the scenario at which the program was stuck
2. What possible execution paths were available from that point.

Solution
--------
Provide a `scenario` function which:

1. Given a group of bthreads and a scenario, checks whether the events
   specified are possible execution paths for the program
2. If not, return (in data) where the program got stuck, and what the
   available execution paths at that point in the program were.