Bugfix: Bid both requests and waits on an event
Problem
-------
If a bid both requests and waits on the same event, then a deadlock
occurs.

It doesn't make sense to do this -- since a request already subscribes
to an event, and a wait adds nothing. Still, we don't want to deadlock.

Solution
---------
Remove any waits that are also requested in the same bid.